+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "concurrently npm:backend:watch npm:frontend:watch",
+    "dev": "concurrently --handle-input npm:backend:watch npm:frontend:watch",
     "backend:watch": "nodemon backend/app.js",
     "frontend:watch": "parcel frontend/index.html"
   },
@@ -15,6 +15,7 @@
   "dependencies": {
     "body-parser": "^1.20.0",
     "concurrently": "^7.1.0",
+    "cors": "^2.8.5",
     "dotenv": "^16.0.0",
     "express": "^4.17.3",
     "hbs": "^4.2.0",
